+++ AzeriteUI/Components/UnitFrames/Units/Party.lua
@@ -33,6 +33,7 @@ end
 local unpack = unpack
 
 -- WoW API
+local UnitGroupRolesAssigned = UnitGroupRolesAssigned
 local UnitIsUnit = UnitIsUnit
 
 -- Addon API
@@ -53,9 +54,454 @@ end
 
 -- Element Callbacks
 --------------------------------------------
+-- Forceupdate health prediction on health updates,
+-- to assure our smoothed elements are properly aligned.
+local Health_PostUpdate = function(element, unit, cur, max)
+	local predict = element.__owner.HealthPrediction
+	if (predict) then
+		predict:ForceUpdate()
+	end
+end
+
+-- Update the health preview color on health color updates.
+local Health_PostUpdateColor = function(element, unit, r, g, b)
+	local preview = element.Preview
+	if (preview) then
+		preview:SetStatusBarColor(r * .7, g * .7, b * .7)
+	end
+end
+
+-- Align our custom health prediction texture
+-- based on the plugin's provided values.
+local HealPredict_PostUpdate = function(element, unit, myIncomingHeal, otherIncomingHeal, absorb, healAbsorb, hasOverAbsorb, hasOverHealAbsorb, curHealth, maxHealth)
+
+	local allIncomingHeal = myIncomingHeal + otherIncomingHeal
+	local allNegativeHeals = healAbsorb
+	local showPrediction, change
+
+	if ((allIncomingHeal > 0) or (allNegativeHeals > 0)) and (maxHealth > 0) then
+		local startPoint = curHealth/maxHealth
+
+		-- Dev switch to test absorbs with normal healing
+		--allIncomingHeal, allNegativeHeals = allNegativeHeals, allIncomingHeal
+
+		-- Hide predictions if the change is very small, or if the unit is at max health.
+		change = (allIncomingHeal - allNegativeHeals)/maxHealth
+		if ((curHealth < maxHealth) and (change > (element.health.predictThreshold or .05))) then
+			local endPoint = startPoint + change
+
+			-- Crop heal prediction overflows
+			if (endPoint > 1) then
+				endPoint = 1
+				change = endPoint - startPoint
+			end
+
+			-- Crop heal absorb overflows
+			if (endPoint < 0) then
+				endPoint = 0
+				change = -startPoint
+			end
+
+			-- This shouldn't happen, but let's do it anyway.
+			if (startPoint ~= endPoint) then
+				showPrediction = true
+			end
+		end
+	end
+
+	if (showPrediction) then
+
+		local preview = element.preview
+		local growth = preview:GetGrowth()
+		local min,max = preview:GetMinMaxValues()
+		local value = preview:GetValue() / max
+		local previewTexture = preview:GetStatusBarTexture()
+		local previewWidth, previewHeight = preview:GetSize()
+		local left, right, top, bottom = preview:GetTexCoord()
+		local isFlipped = preview:IsFlippedHorizontally()
+
+		if (growth == "RIGHT") then
+
+			local texValue, texChange = value, change
+			local rangeH, rangeV
+
+			rangeH = right - left
+			rangeV = bottom - top
+			texChange = change*value
+			texValue = left + value*rangeH
+
+			if (change > 0) then
+				element:ClearAllPoints()
+				element:SetPoint("BOTTOMLEFT", previewTexture, "BOTTOMRIGHT", 0, 0)
+				element:SetSize(change*previewWidth, previewHeight)
+				if (isFlipped) then
+					element:SetTexCoord(texValue + texChange, texValue, top, bottom)
+				else
+					element:SetTexCoord(texValue, texValue + texChange, top, bottom)
+				end
+				element:SetVertexColor(0, .7, 0, .25)
+				element:Show()
+
+			elseif (change < 0) then
+				element:ClearAllPoints()
+				element:SetPoint("BOTTOMRIGHT", previewTexture, "BOTTOMRIGHT", 0, 0)
+				element:SetSize((-change)*previewWidth, previewHeight)
+				if (isFlipped) then
+					element:SetTexCoord(texValue, texValue + texChange, top, bottom)
+				else
+					element:SetTexCoord(texValue + texChange, texValue, top, bottom)
+				end
+				element:SetVertexColor(.5, 0, 0, .75)
+				element:Show()
+
+			else
+				element:Hide()
+			end
+
+		elseif (growth == "LEFT") then
+			local texValue, texChange = value, change
+			local rangeH, rangeV
+			rangeH = right - left
+			rangeV = bottom - top
+			texChange = change*value
+			texValue = left + value*rangeH
+
+			if (change > 0) then
+				element:ClearAllPoints()
+				element:SetPoint("BOTTOMRIGHT", previewTexture, "BOTTOMLEFT", 0, 0)
+				element:SetSize(change*previewWidth, previewHeight)
+				if (isFlipped) then
+					element:SetTexCoord(texValue, texValue + texChange, top, bottom)
+				else
+					element:SetTexCoord(texValue + texChange, texValue, top, bottom)
+				end
+				element:SetVertexColor(0, .7, 0, .25)
+				element:Show()
+
+			elseif (change < 0) then
+				element:ClearAllPoints()
+				element:SetPoint("BOTTOMLEFT", previewTexture, "BOTTOMLEFT", 0, 0)
+				element:SetSize((-change)*previewWidth, previewHeight)
+				if (isFlipped) then
+					element:SetTexCoord(texValue + texChange, texValue, top, bottom)
+				else
+					element:SetTexCoord(texValue, texValue + texChange, top, bottom)
+				end
+				element:SetVertexColor(.5, 0, 0, .75)
+				element:Show()
+
+			else
+				element:Hide()
+			end
+		end
+	else
+		element:Hide()
+	end
+
+	local absorb = element.Absorb
+	if (absorb) then
+		local fraction = absorb/maxHealth
+		if (fraction > .6) then
+			absorb = maxHealth * .6
+		end
+		absorb:SetMinMaxValues(0, maxHealth)
+		absorb:SetValue(absorb)
+	end
+
+end
+
+-- Custom Group Role updater
+local GroupRoleIndicator_Override = function(self, event)
+	local element = self.GroupRoleIndicator
+
+	--[[ Callback: GroupRoleIndicator:PreUpdate()
+	Called before the element has been updated.
+
+	* self - the GroupRoleIndicator element
+	--]]
+	if (element.PreUpdate) then
+		element:PreUpdate()
+	end
+
+	local role = UnitGroupRolesAssigned(self.unit)
+	if (role and element[role]) then
+		element.Icon:SetTexture(element[role])
+		element:Show()
+	else
+		element:Hide()
+	end
+
+	--[[ Callback: GroupRoleIndicator:PostUpdate(role)
+	Called after the element has been updated.
+
+	* self - the GroupRoleIndicator element
+	* role - the role as returned by [UnitGroupRolesAssigned](http://wowprogramming.com/docs/api/UnitGroupRolesAssigned.html)
+	--]]
+	if (element.PostUpdate) then
+		return element:PostUpdate(role)
+	end
+end
+
+-- Update targeting highlight outline
+local TargetHighlight_Update = function(self, event, unit, ...)
+	if (unit and unit ~= self.unit) then return end
+
+	local element = self.TargetHighlight
+	unit = unit or self.unit
+
+	if (UnitIsUnit(unit, "target")) then
+		element:SetVertexColor(unpack(element.colorTarget))
+		element:Show()
+	else
+		element:Hide()
+	end
+
+end
+
+local UnitFrame_PostUpdate = function(self)
+	TargetHighlight_Update(self)
+end
 
 -- Frame Script Handlers
 --------------------------------------------
+local OnEvent = function(self, event, unit, ...)
+	UnitFrame_PostUpdate(self)
+end
 
 UnitStyles["Party"] = function(self, unit, id, ...)
+
+	local db = ns.Config.Party
+
+	self:SetSize(unpack(db.PartySize))
+	self:SetHitRectInsets(unpack(db.PartyHitRectInsets))
+	self:SetFrameLevel(self:GetFrameLevel() + 10)
+
+	-- Overlay for icons and text
+	--------------------------------------------
+	local overlay = CreateFrame("Frame", nil, self)
+	overlay:SetFrameLevel(self:GetFrameLevel() + 7)
+	overlay:SetAllPoints()
+
+	self.Overlay = overlay
+
+	-- Health
+	--------------------------------------------
+	local health = self:CreateBar()
+	health:SetFrameLevel(health:GetFrameLevel() + 2)
+	health:SetPoint(unpack(db.HealthBarPosition))
+	health:SetSize(unpack(db.HealthBarSize))
+	health:SetStatusBarTexture(db.HealthBarTexture)
+	health:SetOrientation(db.HealthBarOrientation)
+	health:SetSparkMap(db.HealthBarSparkMap)
+	health.predictThreshold = .01
+	health.colorDisconnected = true
+	health.colorClass = true
+	health.colorClassPet = true
+	health.colorReaction = true
+	health.colorHealth = true
+
+	self.Health = health
+	self.Health.Override = ns.API.UpdateHealth
+	self.Health.PostUpdate = Health_PostUpdate
+	self.Health.PostUpdateColor = Health_PostUpdateColor
+
+	local healthBackdrop = health:CreateTexture(nil, "BACKGROUND", nil, -1)
+	healthBackdrop:SetPoint(unpack(db.HealthBackdropPosition))
+	healthBackdrop:SetSize(unpack(db.HealthBackdropSize))
+	healthBackdrop:SetTexture(db.HealthBackdropTexture)
+	healthBackdrop:SetVertexColor(unpack(db.HealthBackdropColor))
+
+	self.Health.Backdrop = healthBackdrop
+
+	local healthPreview = self:CreateBar(nil, health)
+	healthPreview:SetAllPoints(health)
+	healthPreview:SetFrameLevel(health:GetFrameLevel() - 1)
+	healthPreview:SetStatusBarTexture(db.HealthBarTexture)
+	healthPreview:SetOrientation(db.HealthBarOrientation)
+	healthPreview:SetSparkTexture("")
+	healthPreview:SetAlpha(.5)
+	healthPreview:DisableSmoothing(true)
+
+	self.Health.Preview = healthPreview
+
+	-- Health Prediction
+	--------------------------------------------
+	local healPredictFrame = CreateFrame("Frame", nil, health)
+	healPredictFrame:SetFrameLevel(health:GetFrameLevel() + 2)
+
+	local healPredict = healPredictFrame:CreateTexture(nil, "OVERLAY", nil, 1)
+	healPredict:SetTexture(db.HealthBarTexture)
+	healPredict.health = health
+	healPredict.preview = healthPreview
+	healPredict.maxOverflow = 1
+
+	self.HealthPrediction = healPredict
+	self.HealthPrediction.PostUpdate = HealPredict_PostUpdate
+
+	-- Cast Overlay
+	--------------------------------------------
+	local castbar = self:CreateBar()
+	castbar:SetAllPoints(health)
+	castbar:SetFrameLevel(self:GetFrameLevel() + 5)
+	castbar:SetSparkMap(db.HealthBarSparkMap)
+	castbar:SetStatusBarTexture(db.HealthBarTexture)
+	castbar:SetStatusBarColor(unpack(db.HealthCastOverlayColor))
+	castbar:DisableSmoothing(true)
+
+	self.Castbar = castbar
+
+	-- Health Value
+	--------------------------------------------
+	local healthValue = health:CreateFontString(nil, "OVERLAY", nil, 1)
+	healthValue:SetPoint(unpack(db.HealthValuePosition))
+	healthValue:SetFontObject(db.HealthValueFont)
+	healthValue:SetTextColor(unpack(db.HealthValueColor))
+	healthValue:SetJustifyH(db.HealthValueJustifyH)
+	healthValue:SetJustifyV(db.HealthValueJustifyV)
+	self:Tag(healthValue, prefix("[*:Health(true)]"))
+
+	self.Health.Value = healthValue
+
+	-- Power
+	--------------------------------------------
+
+
+	-- Absorb Bar (Retail)
+	--------------------------------------------
+	if (ns.IsRetail) then
+		local absorb = self:CreateBar()
+		absorb:SetAllPoints(health)
+		absorb:SetFrameLevel(health:GetFrameLevel() + 3)
+		absorb:SetStatusBarTexture(db.HealthBarTexture)
+		absorb:SetStatusBarColor(unpack(db.HealthAbsorbColor))
+		absorb:SetSparkMap(db.HealthBarSparkMap)
+
+		local orientation
+		if (db.HealthBarOrientation == "UP") then
+			orientation = "DOWN"
+		elseif (db.HealthBarOrientation == "DOWN") then
+			orientation = "UP"
+		elseif (db.HealthBarOrientation == "LEFT") then
+			orientation = "RIGHT"
+		else
+			orientation = "LEFT"
+		end
+		absorb:SetOrientation(orientation)
+
+		self.Health.Absorb = absorb
+	end
+
+	-- Readycheck
+	--------------------------------------------
+	local readyCheckIndicator = overlay:CreateTexture(nil, "OVERLAY", nil, 7)
+	readyCheckIndicator:SetSize(unpack(db.ReadyCheckSize))
+	readyCheckIndicator:SetPoint(unpack(db.ReadyCheckPosition))
+	readyCheckIndicator.readyTexture = db.ReadyCheckReadyTexture
+	readyCheckIndicator.notReadyTexture = db.ReadyCheckNotReadyTexture
+	readyCheckIndicator.waitingTexture = db.ReadyCheckWaitingTexture
+
+	self.ReadyCheckIndicator = ReadyCheckIndicator
+
+	-- Ressurection Indicator
+	--------------------------------------------
+	local resurrectIndicator = overlay:CreateTexture(nil, "OVERLAY", nil, 1)
+	resurrectIndicator:SetSize(unpack(db.ResurrectIndicatorSize))
+	resurrectIndicator:SetPoint(unpack(db.ResurrectIndicatorPosition))
+	resurrectIndicator:SetTexture(ResurrectIndicatorTexture)
+
+	self.ResurrectIndicator = resurrectIndicator
+
+	-- Group Role
+	-----------------------------------------
+    local groupRoleIndicator = CreateFrame("Frame", nil, overlay)
+	groupRoleIndicator:SetSize(unpack(db.GroupRoleSize))
+	groupRoleIndicator:SetPoint(unpack(db.GroupRolePosition))
+	groupRoleIndicator.DAMAGER = db.GroupRoleDPSTexture
+	groupRoleIndicator.HEALER = db.GroupRoleHealerTexture
+	groupRoleIndicator.TANK = db.GroupRoleTankTexture
+
+	local groupRoleBackdrop = groupRoleIndicator:CreateTexture(nil, "BACKGROUND", nil, 1)
+	groupRoleBackdrop:SetSize(unpack(db.GroupRoleBackdropSize))
+	groupRoleBackdrop:SetPoint(unpack(db.GroupRoleBackdropPosition))
+	groupRoleBackdrop:SetVertexColor(unpack(db.GroupRoleBackdropColor))
+
+	groupRoleIndicator.Backdrop = groupRoleBackdrop
+
+	local groupRoleIcon = groupRoleIndicator:CreateTexture(nil, "ARTWORK", nil, 1)
+	groupRoleIcon:SetSize(unpack(db.GroupRoleIconSize))
+	groupRoleIcon:SetPoint(unpack(db.GroupRoleIconPositon))
+
+	groupRoleIndicator.Icon = groupRoleIcon
+
+    self.GroupRoleIndicator = groupRoleIndicator
+	self.GroupRoleIndicator.Override = GroupRoleIndicator_Override
+
+	-- CombatFeedback Text
+	--------------------------------------------
+	local feedbackText = overlay:CreateFontString(nil, "OVERLAY")
+	feedbackText:SetPoint(db.CombatFeedbackPosition[1], self[db.CombatFeedbackAnchorElement], unpack(db.CombatFeedbackPosition))
+	feedbackText:SetFontObject(db.CombatFeedbackFont)
+	feedbackText.feedbackFont = db.CombatFeedbackFont
+	feedbackText.feedbackFontLarge = db.CombatFeedbackFontLarge
+	feedbackText.feedbackFontSmall = db.CombatFeedbackFontSmall
+
+	self.CombatFeedback = feedbackText
+
+	-- Target Highlight
+	--------------------------------------------
+	local targetHighlight = overlay:CreateTexture(nil, "BACKGROUND", nil, -2)
+	targetHighlight:SetPoint(unpack(db.TargetHighlightPosition))
+	targetHighlight:SetSize(unpack(db.TargetHighlightSize))
+	targetHighlight:SetTexture(db.TargetHighlightTexture)
+	targetHighlight.colorTarget = db.TargetHighlightTargetColor
+	targetHighlight.colorFocus = db.TargetHighlightFocusColor
+
+	self.TargetHighlight = targetHighlight
+
+	-- Auras
+	--------------------------------------------
+	local auras = CreateFrame("Frame", nil, self)
+	auras:SetSize(unpack(db.AurasSize))
+	auras:SetPoint(unpack(db.AurasPosition))
+	auras.size = db.AuraSize
+	auras.spacing = db.AuraSpacing
+	auras.numTotal = db.AurasNumTotal
+	auras.disableMouse = db.AurasDisableMouse
+	auras.disableCooldown = db.AurasDisableCooldown
+	auras.onlyShowPlayer = db.AurasOnlyShowPlayer
+	auras.showStealableBuffs = db.AurasShowStealableBuffs
+	auras.initialAnchor = db.AurasInitialAnchor
+	auras["spacing-x"] = db.AurasSpacingX
+	auras["spacing-y"] = db.AurasSpacingY
+	auras["growth-x"] = db.AurasGrowthX
+	auras["growth-y"] = db.AurasGrowthY
+	auras.tooltipAnchor = db.AurasTooltipAnchor
+	auras.sortMethod = db.AurasSortMethod
+	auras.sortDirection = db.AurasSortDirection
+	auras.reanchorIfVisibleChanged = true
+	auras.CreateButton = ns.AuraStyles.CreateButton
+	auras.PostUpdateButton = ns.AuraStyles.TargetPostUpdateButton
+	auras.CustomFilter = ns.AuraFilters.TargetAuraFilter
+	auras.PreSetPosition = ns.AuraSorts.Default -- only in classic
+	auras.SortAuras = ns.AuraSorts.DefaultFunction -- only in retail
+
+	self.Auras = auras
+
+	-- Range Alpha
+	--------------------------------------------
+	self.Range = {
+		insideAlpha = 1,
+		outsideAlpha = db.OutOfRangeAlpha,
+	}
+
+	-- Add a callback for external style overriders
+	self:AddForceUpdate(UnitFrame_PostUpdate)
+
+	-- Textures need an update when frame is displayed.
+	self.PostUpdate = UnitFrame_PostUpdate
+
+	-- Register events to handle additional texture updates.
+	self:RegisterEvent("PLAYER_ENTERING_WORLD", OnEvent, true)
+	self:RegisterEvent("PLAYER_TARGET_CHANGED", OnEvent, true)
+
 end