+++ CHANGELOG.md
@@ -4,5 +4,9 @@ All notable changes to this project will be documented in this file. Be aware th
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] 2022-11-27
+### Added
+- Added party frames.
+
 ## [4.0.0-Alpha] 2022-11-25
 - First public alpha.